+++ NEWS.md
@@ -3,7 +3,7 @@
   - Bioconductor packages are used conditionally
   - Package website with `pkgdown`
   - Open Tree of Life cached chronogram database not yet updated
-  - get_otol_chronograms was updated and renamed to get_opentree_chronograms
+  - get_opentree_chronograms was updated and renamed to get_opentree_chronograms
   - documentation for all functions was expanded
   - examples and function files are written to tempdir()